% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{employees_by_workplace}
\alias{employees_by_workplace}
\alias{employees_by_residence}
\title{Data about employees}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4410198 rows and 6 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4410198 rows and 6 columns.
}
\usage{
employees_by_workplace

employees_by_residence
}
\description{
Data about employees
}
\keyword{datasets}
