% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jobs}
\alias{jobs}
\title{Data about jobs}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 816060 rows and 5 columns.
}
\usage{
jobs
}
\description{
Data about jobs
}
\keyword{datasets}
