% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{occupational_group_codes}
\alias{occupational_group_codes}
\alias{region_codes}
\title{Occupational group and region codes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 145 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 401 rows and 2 columns.
}
\usage{
occupational_group_codes

region_codes
}
\description{
Occupational group and region codes
}
\keyword{datasets}
