% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unemployed_total}
\alias{unemployed_total}
\alias{unemployed_foreigners}
\title{Data about unemployed persons}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 814030 rows and 5 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 814030 rows and 5 columns.
}
\usage{
unemployed_total

unemployed_foreigners
}
\description{
Data about unemployed persons
}
\keyword{datasets}
